Adjusted Sample CSV to match script
Now that the script is simplified and no longer handles answer regex, the header row of the CSV only includes questions. Also added a greeting utterance to each conversation sequence so that the actual user intended utterance is passed in the same conversation.  This allows dialog to be more properly triggered.