feat(Conversation): Add script to generate logs from CSV file (#14) (#15)
Based on the PR changes requested and German's rewritten script.  German made a bunch of improvements, including supporting Python 2 and using the Conversation Python SDK.  I tweaked German's version in small ways including
- added more checks on input parameters. Made url optional with a default.
- removed the CSV answer regex and intents handling. Their origin was for testing while this script is focused on generation only.

* Adjusted Sample CSV to match the script

Now that the script is simplified and no longer handles answer regex, the header row of the CSV only includes questions. Also added a greeting utterance to each conversation sequence so that the actual user intended utterance is passed in the same conversation.  This allows the dialog to be more properly triggered.

* undo putting CSV in the wrong place

* Change CSV to match the script

Now that the script is simplified and no longer handles answer regex, the header row of the CSV only includes questions. Also added a greeting utterance to each conversation sequence so that the actual user intended utterance is passed in the same conversation.  This allows the dialog to be more properly triggered.